float -> int

+++ inc/cfs_visibility.h
@@ -4,11 +4,11 @@ class Cfs_visibility {
 public:
 
     int schedule_id;
-    float begin;
-    float end;
+    int begin;
+    int end;
 
-    Cfs_visibility(int schedule_id, float begin, float end)
+    Cfs_visibility(int schedule_id, int begin, int end)
         : schedule_id(schedule_id), begin(begin), end(end) {}
-    bool is_active_at_time(float timestamp) const ;
+    bool is_active_at_time(int timestamp) const ;
 
 };
