+++ src/cfs_visibility.cc
@@ -1,5 +1,5 @@
 #include <cfs_visibility.h>
 
-bool Cfs_visibility::is_active_at_time(float timestamp) const {
+bool Cfs_visibility::is_active_at_time(int timestamp) const {
     return (this->begin <= timestamp && this->end > timestamp);
 }
